fix: replace unsigned int types with signed to fix OpenCode schema warnings
Change u16/u32/u64/usize to i16/i32/i64 in all JSON Schema struct fields
to eliminate 'unknown format' warnings in OpenCode. All 40 tests pass.

+++ src/tools/disk.rs
@@ -20,7 +20,7 @@ pub struct DiskInfo {
 #[derive(Debug, Serialize, JsonSchema)]
 pub struct DiskInfoOutput {
     pub disks: Vec<DiskInfo>,
-    pub total_disks: usize,
+    pub total_disks: i32,
 }
 
 pub fn disk_info_inner(_args: DiskInfoArgs) -> DiskInfoOutput {
@@ -77,7 +77,7 @@ pub fn disk_info_inner(_args: DiskInfoArgs) -> DiskInfoOutput {
         Vec::new()
     };
 
-    let total_disks = disks.len();
+    let total_disks = disks.len() as i32;
     DiskInfoOutput {
         disks,
         total_disks,
